chatbot-backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
qdrant-client
langchain
tiktoken
openai
sentence-transformers
